1. Implemented Matt's branch 2. Deleted extra directory layer of "djangoproject". This allows urls.py to take imports from views.py 3. Renamed lower directory to "technical"
