--- .idea/Decryptors.iml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<module version="4">
-  <component name="PyDocumentationSettings">
-    <option name="format" value="PLAIN" />
-    <option name="myDocStringFormat" value="Plain" />
-  </component>
-<<<<<<<<< Temporary merge branch 1
-  <component name="TemplatesService">
-    <option name="TEMPLATE_FOLDERS">
-      <list>
-        <option value="$MODULE_DIR$/djangoproject/templates" />
-      </list>
-    </option>
-  </component>
-=========
->>>>>>>>> Temporary merge branch 2
-</module>
