+++ technical/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for djangoproject project.
+Django settings for technical project.
 
 Generated by 'django-admin startproject' using Django 4.2.
 
@@ -50,7 +50,7 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'djangoproject.urls'
+ROOT_URLCONF = 'technical.urls'
 
 TEMPLATES = [
     {
@@ -69,7 +69,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'djangoproject.wsgi.application'
+WSGI_APPLICATION = 'technical.wsgi.application'
 
 
 # Database
