+++ templates/courses.html
